Light components and Scene tabbing background
AmbientLight, DirectionalLight, PointLight, SpotLight are components that can be added to affect the lighting of the object and childrens. Missing falloff.

LightSourceComponent is the base class of the lights. It contain the functions enable and disable to control the light.

The views for thoses components are also created.

Optimised getGameObjectAt to make it faster.

Added generateEmpty() in demoscene.cpp.

When creating a rootView, it new delete the old one. This was made for the next feature :
Ctrl+N to create a new scene
Ctrl+W to delete current selected scene (if last scene, then create a new empty one)
Ctrl+Tab to navigate to the next scene
Ctrl+Shift+Tab to navigate to preview scene